requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.5
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
numpy==1.22.4
pandas==1.4.2
pipreqs==0.4.11
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==2.10.6
requests==2.28.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
Unipath==1.1
urllib3==1.26.9
whitenoise==6.2.0
yarg==0.1.9
zope.interface==5.4.0
